refactor: Remove example rgbpp utils
