+++ examples/rgbpp/package.json
@@ -13,10 +13,7 @@
     "@exact-realty/multipart-parser": "^1.0.13",
     "@nervosnetwork/ckb-sdk-utils": "^0.109.1",
     "@spore-sdk/core": "^0.2.0-beta.6",
-    "rgbpp": "workspace:^",
-    "@rgbpp-sdk/service": "workspace:^",
-    "@rgbpp-sdk/ckb": "workspace:^",
-    "@rgbpp-sdk/btc": "workspace:^"
+    "rgbpp": "workspace:^"
   },
   "devDependencies": {
     "@types/node": "^20.11.28",
