+++ examples/rgbpp/spore/5-leap-spore-to-ckb.ts
@@ -1,8 +1,51 @@
-import { buildRgbppLockArgs } from '@rgbpp-sdk/ckb';
-import { leapSporeToCkb, btcService, SporeLeapParams } from '../utils';
+import { buildRgbppLockArgs } from 'rgbpp/ckb';
+import { genLeapSporeFromBtcToCkbVirtualTx, sendRgbppUtxos } from 'rgbpp';
+import { getSporeTypeScript, Hex } from 'rgbpp/ckb';
+import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
+import { isMainnet, collector, btcAddress, btcDataSource, btcKeyPair, btcService } from '../env';
 
-const leapSpore = async (params: SporeLeapParams) => {
-  const btcTxId = await leapSporeToCkb(params);
+interface SporeLeapParams {
+  sporeRgbppLockArgs: Hex;
+  toCkbAddress: string;
+  sporeTypeArgs: Hex;
+}
+
+const leapSporeFromBtcToCkb = async ({ sporeRgbppLockArgs, toCkbAddress, sporeTypeArgs }: SporeLeapParams) => {
+  const sporeTypeBytes = serializeScript({
+    ...getSporeTypeScript(isMainnet),
+    args: sporeTypeArgs,
+  });
+
+  const ckbVirtualTxResult = await genLeapSporeFromBtcToCkbVirtualTx({
+    collector,
+    sporeRgbppLockArgs,
+    sporeTypeBytes,
+    toCkbAddress,
+    isMainnet,
+  });
+
+  const { commitment, ckbRawTx } = ckbVirtualTxResult;
+
+  // Send BTC tx
+  const psbt = await sendRgbppUtxos({
+    ckbVirtualTx: ckbRawTx,
+    commitment,
+    tos: [btcAddress!],
+    ckbCollector: collector,
+    from: btcAddress!,
+    source: btcDataSource,
+    feeRate: 30,
+  });
+  psbt.signAllInputs(btcKeyPair);
+  psbt.finalizeAllInputs();
+
+  const btcTx = psbt.extractTransaction();
+  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
+
+  console.log('BTC TxId: ', btcTxId);
+
+  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
+
   try {
     const interval = setInterval(async () => {
       const { state, failedReason } = await btcService.getRgbppTransactionState(btcTxId);
@@ -24,7 +67,7 @@ const leapSpore = async (params: SporeLeapParams) => {
 
 // Use your real BTC UTXO information on the BTC Testnet
 // rgbppLockArgs: outIndexU32 + btcTxId
-leapSpore({
+leapSporeFromBtcToCkb({
   sporeRgbppLockArgs: buildRgbppLockArgs(3, 'd8a31796fbd42c546f6b22014b9b82b16586ce1df81b0e7ca9a552cdc492a0af'),
   toCkbAddress: 'ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsq0e4xk4rmg5jdkn8aams492a7jlg73ue0gc0ddfj',
   sporeTypeArgs: '0x42898ea77062256f46e8f1b861d526ae47810ecc51ab50477945d5fa90452706',
