+++ examples/rgbpp/spore/6-unlock-btc-time-cell.ts
@@ -1,6 +1,6 @@
 import { buildSporeBtcTimeCellsSpentTx, signBtcTimeCellSpentTx } from 'rgbpp';
-import { CKB_PRIVATE_KEY, btcService, ckbAddress, collector, isMainnet } from '../utils';
-import { sendCkbTx, getBtcTimeLockScript } from '@rgbpp-sdk/ckb';
+import { CKB_PRIVATE_KEY, btcService, ckbAddress, collector, isMainnet } from '../env';
+import { sendCkbTx, getBtcTimeLockScript } from 'rgbpp/ckb';
 
 // Warning: Wait at least 6 BTC confirmation blocks to spend the BTC time cells after 4-btc-jump-ckb.ts
 const unlockSporeBtcTimeCell = async ({ btcTimeCellArgs }: { btcTimeCellArgs: string }) => {
