+++ examples/rgbpp/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "target": "ES2015",
     "lib": ["dom", "dom.iterable", "esnext"],
-    "module": "CommonJS",
+    "module": "NodeNext",
     "composite": false,
     "resolveJsonModule": true,
     "strictNullChecks": true,
@@ -14,7 +14,7 @@
     "forceConsistentCasingInFileNames": true,
     "inlineSources": false,
     "isolatedModules": true,
-    "moduleResolution": "node",
+    "moduleResolution": "NodeNext",
     "noUnusedLocals": false,
     "noUnusedParameters": false,
     "preserveWatchOutput": true,
