--- examples/rgbpp/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './spore';
-export * from './xudt';
-export * from './env';
