--- examples/rgbpp/utils/xudt.ts
@@ -1,103 +1,0 @@
-import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
-import { getXudtTypeScript } from '@rgbpp-sdk/ckb';
-import { genBtcTransferCkbVirtualTx, genBtcJumpCkbVirtualTx, sendRgbppUtxos } from 'rgbpp';
-import { isMainnet, collector, btcAddress, btcKeyPair, btcService, btcDataSource } from './env';
-
-export interface RgbppTransferParams {
-  rgbppLockArgsList: string[];
-  toBtcAddress: string;
-  xudtTypeArgs: string;
-  transferAmount: bigint;
-}
-export const transferRgbppOnBtc = async ({
-  rgbppLockArgsList,
-  toBtcAddress,
-  xudtTypeArgs,
-  transferAmount,
-}: RgbppTransferParams): Promise<string> => {
-  const xudtType: CKBComponents.Script = {
-    ...getXudtTypeScript(isMainnet),
-    args: xudtTypeArgs,
-  };
-
-  const ckbVirtualTxResult = await genBtcTransferCkbVirtualTx({
-    collector,
-    rgbppLockArgsList,
-    xudtTypeBytes: serializeScript(xudtType),
-    transferAmount,
-    isMainnet,
-  });
-
-  const { commitment, ckbRawTx } = ckbVirtualTxResult;
-
-  // Send BTC tx
-  const psbt = await sendRgbppUtxos({
-    ckbVirtualTx: ckbRawTx,
-    commitment,
-    tos: [toBtcAddress],
-    ckbCollector: collector,
-    from: btcAddress!,
-    source: btcDataSource,
-  });
-  psbt.signAllInputs(btcKeyPair);
-  psbt.finalizeAllInputs();
-
-  const btcTx = psbt.extractTransaction();
-  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
-
-  console.log('BTC TxId: ', btcTxId);
-
-  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
-
-  return btcTxId;
-};
-
-export interface LeapToCkbParams {
-  rgbppLockArgsList: string[];
-  toCkbAddress: string;
-  xudtTypeArgs: string;
-  transferAmount: bigint;
-}
-export const leapFromBtcToCkb = async ({
-  rgbppLockArgsList,
-  toCkbAddress,
-  xudtTypeArgs,
-  transferAmount,
-}: LeapToCkbParams): Promise<string> => {
-  const xudtType: CKBComponents.Script = {
-    ...getXudtTypeScript(isMainnet),
-    args: xudtTypeArgs,
-  };
-
-  const ckbVirtualTxResult = await genBtcJumpCkbVirtualTx({
-    collector,
-    rgbppLockArgsList,
-    xudtTypeBytes: serializeScript(xudtType),
-    transferAmount,
-    toCkbAddress,
-    isMainnet,
-  });
-
-  const { commitment, ckbRawTx } = ckbVirtualTxResult;
-
-  // Send BTC tx
-  const psbt = await sendRgbppUtxos({
-    ckbVirtualTx: ckbRawTx,
-    commitment,
-    tos: [btcAddress!],
-    ckbCollector: collector,
-    from: btcAddress!,
-    source: btcDataSource,
-  });
-  psbt.signAllInputs(btcKeyPair);
-  psbt.finalizeAllInputs();
-
-  const btcTx = psbt.extractTransaction();
-  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
-
-  console.log('BTC TxId: ', btcTxId);
-
-  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
-
-  return btcTxId;
-};
