+++ examples/rgbpp/xudt/3-btc-leap-ckb.ts
@@ -1,8 +1,51 @@
-import { buildRgbppLockArgs } from '@rgbpp-sdk/ckb';
-import { LeapToCkbParams, btcService, leapFromBtcToCkb } from '../utils';
+import { buildRgbppLockArgs, getXudtTypeScript } from 'rgbpp/ckb';
+import { serializeScript } from '@nervosnetwork/ckb-sdk-utils';
+import { genBtcJumpCkbVirtualTx, sendRgbppUtxos } from 'rgbpp';
+import { isMainnet, collector, btcAddress, btcKeyPair, btcService, btcDataSource } from '../env';
 
-const leapToCKB = async (params: LeapToCkbParams) => {
-  const btcTxId = await leapFromBtcToCkb(params);
+interface LeapToCkbParams {
+  rgbppLockArgsList: string[];
+  toCkbAddress: string;
+  xudtTypeArgs: string;
+  transferAmount: bigint;
+}
+
+const leapFromBtcToCKB = async ({ rgbppLockArgsList, toCkbAddress, xudtTypeArgs, transferAmount }: LeapToCkbParams) => {
+  const xudtType: CKBComponents.Script = {
+    ...getXudtTypeScript(isMainnet),
+    args: xudtTypeArgs,
+  };
+
+  const ckbVirtualTxResult = await genBtcJumpCkbVirtualTx({
+    collector,
+    rgbppLockArgsList,
+    xudtTypeBytes: serializeScript(xudtType),
+    transferAmount,
+    toCkbAddress,
+    isMainnet,
+  });
+
+  const { commitment, ckbRawTx } = ckbVirtualTxResult;
+
+  // Send BTC tx
+  const psbt = await sendRgbppUtxos({
+    ckbVirtualTx: ckbRawTx,
+    commitment,
+    tos: [btcAddress!],
+    ckbCollector: collector,
+    from: btcAddress!,
+    source: btcDataSource,
+  });
+  psbt.signAllInputs(btcKeyPair);
+  psbt.finalizeAllInputs();
+
+  const btcTx = psbt.extractTransaction();
+  const { txid: btcTxId } = await btcService.sendBtcTransaction(btcTx.toHex());
+
+  console.log('BTC TxId: ', btcTxId);
+
+  await btcService.sendRgbppCkbTransaction({ btc_txid: btcTxId, ckb_virtual_result: ckbVirtualTxResult });
+
   try {
     const interval = setInterval(async () => {
       const { state, failedReason } = await btcService.getRgbppTransactionState(btcTxId);
@@ -23,7 +66,7 @@ const leapToCKB = async (params: LeapToCkbParams) => {
 };
 
 // rgbppLockArgs: outIndexU32 + btcTxId
-leapToCKB({
+leapFromBtcToCKB({
   rgbppLockArgsList: [buildRgbppLockArgs(1, '6edd4b9327506fab09fb9a0f5e5f35136a6a94bd4c9dd79af04921618fa6c800')],
   toCkbAddress: 'ckt1qrfrwcdnvssswdwpn3s9v8fp87emat306ctjwsm3nmlkjg8qyza2cqgqq9kxr7vy7yknezj0vj0xptx6thk6pwyr0sxamv6q',
   xudtTypeArgs: '0x1ba116c119d1cfd98a53e9d1a615cf2af2bb87d95515c9d217d367054cfc696b',
