+++ examples/xudt/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@nervosnetwork/ckb-sdk-utils": "^0.109.1",
-    "@rgbpp-sdk/ckb": "workspace:^"
+    "rgbpp": "workspace:^"
   },
   "devDependencies": {
     "dotenv": "^16.4.5",
